feat(http): Implement soft and hard TTL for all cache providers
