+++ lib/util/http/cache/repository-http-cache-provider.spec.ts
@@ -1,6 +1,5 @@
 import { Http } from '..';
 import * as httpMock from '../../../../test/http-mock';
-import { logger } from '../../../../test/util';
 import { resetCache } from '../../cache/repository';
 import { repoCacheProvider } from './repository-http-cache-provider';
 
@@ -57,19 +56,6 @@ describe('util/http/cache/repository-http-cache-provider', () => {
       body: { msg: 'Hello, world!' },
       authorization: false,
     });
-  });
-
-  it('reports if cache could not be persisted', async () => {
-    httpMock
-      .scope('https://example.com')
-      .get('/foo/bar')
-      .reply(200, { msg: 'Hello, world!' });
-
-    await http.getJsonUnchecked('https://example.com/foo/bar');
-
-    expect(logger.logger.debug).toHaveBeenCalledWith(
-      'http cache: failed to persist cache for https://example.com/foo/bar',
-    );
   });
 
   it('handles abrupt cache reset', async () => {
