+++ lib/util/http/cache/repository-http-cache-provider.ts
@@ -3,6 +3,9 @@ import { AbstractHttpCacheProvider } from './abstract-http-cache-provider';
 import type { HttpCache } from './schema';
 
 export class RepositoryHttpCacheProvider extends AbstractHttpCacheProvider {
+  protected override softTtlMinutes = 0;
+  protected override hardTtlMinutes = 365 * 24 * 60;
+
   override load(url: string): Promise<unknown> {
     const cache = getCache();
     cache.httpCache ??= {};
